+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## v0.0.3
 - Added Wand of The Archer
 - Added Wand of Celebration
+- Added Space Time Amalgamator (Creates Liquid Space Time)
 - Added crafting recipes for Unholy Silencer and Space Time Evaporator
 
 ---